# api/swagger/swagger.yaml
consumes:
- application/json
definitions:
  Crudwebsocket-Error:
    description: Crudwebsocket-Error media type (default view)
    example:
      code: UNAUTHORIZED
      msg: Unauthorized access
    properties:
      code:
        description: Message ID
        example: UNAUTHORIZED
        type: string
      msg:
        description: Localized message
        example: Unauthorized access
        type: string
    required:
    - code
    - msg
    title: 'Mediatype identifier: application/vnd.crudwebsocket-error+json; view=default'
    type: object
  Crudwebsocket-Version:
    description: Crudwebsocket-Version media type (default view)
    example:
      git: "000000"
      version: "1.0"
    properties:
      git:
        description: Git commit hash
        example: "000000"
        type: string
      version:
        description: Application version
        example: "1.0"
        type: string
    required:
    - version
    title: 'Mediatype identifier: application/vnd.crudwebsocket-version+json; view=default'
    type: object
  error:
    description: Error response media type (default view)
    example:
      code: invalid_value
      detail: Value of ID must be an integer
      id: 3F1FKVRR
      meta:
        timestamp: 1458609066
      status: "400"
    properties:
      code:
        description: an application-specific error code, expressed as a string value.
        example: invalid_value
        type: string
      detail:
        description: a human-readable explanation specific to this occurrence of the
          problem.
        example: Value of ID must be an integer
        type: string
      id:
        description: a unique identifier for this particular occurrence of the problem.
        example: 3F1FKVRR
        type: string
      meta:
        additionalProperties: true
        description: a meta object containing non-standard meta-information about
          the error.
        example:
          timestamp: 1458609066
        type: object
      status:
        description: the HTTP status code applicable to this problem, expressed as
          a string value.
        example: "400"
        type: string
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
info:
  description: API for CRUD WebSocket
  title: CRUD WebSocket API
  version: "1.0"
paths:
  /crudws/create:
    post:
      operationId: crud#create
      produces:
      - application/json
      - application/vnd.crudwebsocket-error+json
      responses:
        "200":
          description: OK
        "204":
          description: No Content
        "400":
          description: BadRequest is returned if input object is missing required
            attributes or their values are out of range.
          schema:
            $ref: '#/definitions/Crudwebsocket-Error'
        "401":
          description: Unauthorized is returned when user request does not contain
            authentication token or authentication is invalid. The response must include
            a valid "WWW-Authenticate" header.
          headers:
            WWW-Authenticate:
              default: Bearer
              description: https://tools.ietf.org/html/rfc7235
              type: string
          schema:
            $ref: '#/definitions/Crudwebsocket-Error'
        "403":
          description: Forbidden is returned when user is not authorized to perform
            an action.
          schema:
            $ref: '#/definitions/Crudwebsocket-Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Crudwebsocket-Error'
      schemes:
      - https
      summary: Get data
      tags:
      - crud
  /crudws/delete/{id}/{idchat}:
    get:
      operationId: crud#delete
      parameters:
      - description: id of users
        in: path
        name: id
        required: true
        type: string
      - description: idchat of chats
        in: path
        name: idchat
        required: true
        type: string
      produces:
      - application/json
      - application/vnd.crudwebsocket-error+json
      responses:
        "200":
          description: OK
        "204":
          description: No Content
        "400":
          description: BadRequest is returned if input object is missing required
            attributes or their values are out of range.
          schema:
            $ref: '#/definitions/Crudwebsocket-Error'
        "401":
          description: Unauthorized is returned when user request does not contain
            authentication token or authentication is invalid. The response must include
            a valid "WWW-Authenticate" header.
          headers:
            WWW-Authenticate:
              default: Bearer
              description: https://tools.ietf.org/html/rfc7235
              type: string
          schema:
            $ref: '#/definitions/Crudwebsocket-Error'
        "403":
          description: Forbidden is returned when user is not authorized to perform
            an action.
          schema:
            $ref: '#/definitions/Crudwebsocket-Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Crudwebsocket-Error'
      schemes:
      - https
      summary: Get data
      tags:
      - crud
  /crudws/select/{id}:
    get:
      operationId: crud#select
      parameters:
      - description: id of items
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      - application/vnd.crudwebsocket-error+json
      responses:
        "200":
          description: OK
        "204":
          description: No Content
        "400":
          description: BadRequest is returned if input object is missing required
            attributes or their values are out of range.
          schema:
            $ref: '#/definitions/Crudwebsocket-Error'
        "401":
          description: Unauthorized is returned when user request does not contain
            authentication token or authentication is invalid. The response must include
            a valid "WWW-Authenticate" header.
          headers:
            WWW-Authenticate:
              default: Bearer
              description: https://tools.ietf.org/html/rfc7235
              type: string
          schema:
            $ref: '#/definitions/Crudwebsocket-Error'
        "403":
          description: Forbidden is returned when user is not authorized to perform
            an action.
          schema:
            $ref: '#/definitions/Crudwebsocket-Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Crudwebsocket-Error'
      schemes:
      - https
      summary: Get data
      tags:
      - crud
  /crudws/update/{id}:
    post:
      operationId: crud#update
      parameters:
      - description: id of items
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      - application/vnd.crudwebsocket-error+json
      responses:
        "200":
          description: OK
        "204":
          description: No Content
        "400":
          description: BadRequest is returned if input object is missing required
            attributes or their values are out of range.
          schema:
            $ref: '#/definitions/Crudwebsocket-Error'
        "401":
          description: Unauthorized is returned when user request does not contain
            authentication token or authentication is invalid. The response must include
            a valid "WWW-Authenticate" header.
          headers:
            WWW-Authenticate:
              default: Bearer
              description: https://tools.ietf.org/html/rfc7235
              type: string
          schema:
            $ref: '#/definitions/Crudwebsocket-Error'
        "403":
          description: Forbidden is returned when user is not authorized to perform
            an action.
          schema:
            $ref: '#/definitions/Crudwebsocket-Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Crudwebsocket-Error'
      schemes:
      - https
      summary: Get data
      tags:
      - crud
  /crudws/version:
    get:
      operationId: version#version
      produces:
      - application/vnd.crudwebsocket-version+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Crudwebsocket-Version'
      schemes:
      - https
      summary: Return application's version and commit hash
      tags:
      - version
  /css/{filepath}:
    get:
      operationId: public#/css/*filepath
      parameters:
      - description: Relative file path
        in: path
        name: filepath
        required: true
        type: string
      responses:
        "200":
          description: File downloaded
          schema:
            type: file
        "404":
          description: File not found
          schema:
            $ref: '#/definitions/error'
      schemes:
      - https
      summary: Download public/css/
  /index:
    get:
      operationId: public#/index
      responses:
        "200":
          description: File downloaded
          schema:
            type: file
      schemes:
      - https
      summary: Download public/html/index.html
  /js/{filepath}:
    get:
      operationId: public#/js/*filepath
      parameters:
      - description: Relative file path
        in: path
        name: filepath
        required: true
        type: string
      responses:
        "200":
          description: File downloaded
          schema:
            type: file
        "404":
          description: File not found
          schema:
            $ref: '#/definitions/error'
      schemes:
      - https
      summary: Download public/js/
  /swagger.json:
    get:
      operationId: swagger#/swagger.json
      responses:
        "200":
          description: File downloaded
          schema:
            type: file
      schemes:
      - https
      summary: Download api/swagger/swagger.json
produces:
- application/json
responses:
  BadRequest:
    description: BadRequest is returned if input object is missing required attributes
      or their values are out of range.
    schema:
      $ref: '#/definitions/Crudwebsocket-Error'
  Forbidden:
    description: Forbidden is returned when user is not authorized to perform an action.
    schema:
      $ref: '#/definitions/Crudwebsocket-Error'
  NoContent:
    description: No Content
  Unauthorized:
    description: Unauthorized is returned when user request does not contain authentication
      token or authentication is invalid. The response must include a valid "WWW-Authenticate"
      header.
    headers:
      WWW-Authenticate:
        default: Bearer
        description: https://tools.ietf.org/html/rfc7235
        type: string
    schema:
      $ref: '#/definitions/Crudwebsocket-Error'
schemes:
- https
swagger: "2.0"